+++ components/bpmn-q/modeler-component/extensions/data-extension/Constants.js
@@ -1,0 +1,13 @@
+// Type names of the data flow extension elements
+export const DATA_MAP_OBJECT = 'dataflow:DataMapObject';
+export const DATA_STORE_MAP = 'dataflow:DataStoreMap';
+export const TRANSFORMATION_TASK= 'dataflow:TransformationTask';
+export const TRANSFORMATION_ASSOCIATION = 'dataflow:TransformationAssociation';
+export const INPUT_TRANSFORMATION_ASSOCIATION = 'dataflow:InputTransformationAssociation';
+export const OUTPUT_TRANSFORMATION_ASSOCIATION = 'dataflow:OutputTransformationAssociation';
+
+// Property names of the data flow extension elements
+export const CONTENT = 'content';
+export const DETAILS = 'details';
+export const PARAMETERS = 'parameters';
+export const EXPRESSIONS = 'expressions';
